Refactoring CLI to be much more modular and pathe way for use as node module + CLI
