+++ package.json
@@ -33,6 +33,7 @@
     "browserify": "^13.1.0",
     "chai": "^3.5.0",
     "chromedriver": "^2.26.1",
+    "clear-require": "^2.0.0",
     "cookie-parser": "^1.4.3",
     "eslint-config-google": "^0.7.0",
     "express": "^4.14.0",
