+++ packages/sw-cli/src/cli/bin.js
