+++ packages/sw-cli/src/index.js
@@ -1,1 +1,6 @@
 // This will be the logic that powers both module and CLI
+const generateSW = require('./lib/generate-sw');
+
+module.exports = {
+  generateSW,
+};
