+++ packages/sw-cli/src/lib/log-helper.js
@@ -23,11 +23,14 @@ const chalk = require('chalk');
  * Log helper is just a wrapper around the console to print slightly
  * nicer / colored messages and could be extended to filter based on log
  * level.
+ * @private
  */
 class LogHelper {
   /**
    * Print a warning message to the console (Colored yellow).
    * @param {string} msg Message to print to the console
+   * @param {Object} [runtimeInfo] Any information available at runtime that
+   * should be printed to the console.
    */
   warn(msg, runtimeInfo) {
     let loggedMsg = chalk.yellow(msg);
@@ -40,6 +43,7 @@ class LogHelper {
   /**
    * Print an error message to the console (Colored red).
    * @param {string} msg Message to print to the console
+   * @param {Error} [err] An optional error to print to the console.
    */
   error(msg, err) {
     console.error(chalk.red(msg));
